Get Cart

+++ src/BoundedContext/CartItems/Application/Delete/CartItemDeleter.php
@@ -8,6 +8,7 @@ namespace Spfc\BoundedContext\CartItems\Application\Delete;
 use Spfc\BoundedContext\CartItems\Domain\CartItemId;
 use Spfc\BoundedContext\CartItems\Domain\CartItemNotExist;
 use Spfc\BoundedContext\CartItems\Domain\CartItemRepository;
+use Spfc\BoundedContext\Carts\Application\Find\CartFinder;
 use Spfc\BoundedContext\Carts\Application\Update\CartTotalsDecrementer;
 use function Lambdish\Phunctional\apply;
 
@@ -18,11 +19,13 @@ final class CartItemDeleter
 
     /**
      * @param CartItemRepository $repository
+     * @param CartFinder $cartFinder
      * @param CartTotalsDecrementer $cartTotalsDecrementer
      */
-    public function __construct(CartItemRepository $repository, CartTotalsDecrementer $cartTotalsDecrementer)
+    public function __construct(CartItemRepository $repository, CartFinder $cartFinder, CartTotalsDecrementer $cartTotalsDecrementer)
     {
         $this->repository = $repository;
+        $this->cartFinder = $cartFinder;
         $this->cartTotalsDecrementer = $cartTotalsDecrementer;
     }
 
@@ -40,8 +43,16 @@ final class CartItemDeleter
             throw new CartItemNotExist($id);
         }
 
+        $cart = $this->cartFinder->__invoke($cartItem->cartId()->value());
+
+        if($cart->payed()->value()) {
+            throw new \InvalidArgumentException(
+                sprintf('<%s> cart is already payed.', $cartItem->cartId())
+            );
+        }
+
         $this->repository->delete($id);
 
-        apply($this->cartTotalsDecrementer, [$cartItem->cartId(), $cartItem->price()->value()]);
+        apply($this->cartTotalsDecrementer, [$cartItem->cartId()->value(), $cartItem->price()->value()]);
     }
 }
